+++ vim/dotfiles/.config/nvim/lua/plugins/bracketpair.lua
@@ -1,4 +1,4 @@
-local pelyib = vim.g.pelyib.pluginconfig
+local pluginconf = require('pelyib.pluginconf').config.patched
 
 return vim.tbl_deep_extend(
     "force",
@@ -9,5 +9,5 @@ return vim.tbl_deep_extend(
         config = function ()
         end
     },
-    pelyib.bracketpair or {}
+    pluginconf.bracketpair or {}
 )