+++ vim/dotfiles/.config/nvim/lua/plugins/devicons.lua
@@ -1,4 +1,4 @@
-local pelyib = vim.g.pelyib.pluginconfig
+local pluginconf = require('pelyib.pluginconf').config.patched
 
 return vim.tbl_deep_extend(
     "force",
@@ -8,5 +8,5 @@ return vim.tbl_deep_extend(
         lazy = false,
         commit = "cff25ce621e6d15fae0b0bfe38c00be50ce38468",
     },
-    pelyib.config.devicons
+    pluginconf.devicons or {}
 )