+++ vim/dotfiles/.config/nvim/lua/plugins/lspzero.lua
@@ -1,4 +1,4 @@
-local pelyib = vim.g.pelyib.pluginconfig
+local pluginconf = require('pelyib.pluginconf').config.patched
 
 return vim.tbl_deep_extend(
     "force",
@@ -146,5 +146,5 @@ return vim.tbl_deep_extend(
             },
         }
     },
-    pelyib.config.lspzero
+    pluginconf.lspzero or {}
 )