+++ vim/dotfiles/.config/nvim/lua/plugins/neotree.lua
@@ -1,4 +1,4 @@
-local pelyib = vim.g.pelyib.pluginconfig
+local pluginconf = require("pelyib.pluginconf").config.patched
 
 return vim.tbl_deep_extend(
     "force",
@@ -38,5 +38,5 @@ return vim.tbl_deep_extend(
             vim.keymap.set('n', 'fb', ':NeoTreeFloatToggle<CR>')
         end
     },
-    pelyib.config.neotree
+    pluginconf.neotree or {}
 )