+++ vim/dotfiles/.config/nvim/lua/plugins/phpactor.lua
@@ -1,4 +1,4 @@
-local pelyib = vim.g.pelyib.pluginconfig
+local pluginconf = require('pelyib.pluginconf').config.patched
 
 return vim.tbl_deep_extend(
     "force",
@@ -12,10 +12,10 @@ return vim.tbl_deep_extend(
             require("pelyib.shell-runner")(
                 {
                     "docker", "run" , "--rm", "--interactive",
-                        "--env", "COMPOSER_HOME",
-                        "--volume", composerHome .. ":" .. composerHome,
-                        "--volume", vim.fn.stdpath("data") .. "/lazy/phpactor:/app",
-                        "composer:2.7.7", "composer", "install", "--no-dev", "--ignore-platform-reqs", "--no-interaction", "--no-progress", "--no-suggest", "--prefer-dist", "--optimize-autoloader"
+                    "--env", "COMPOSER_HOME",
+                    "--volume", composerHome .. ":" .. composerHome,
+                    "--volume", vim.fn.stdpath("data") .. "/lazy/phpactor:/app",
+                    "composer:2.7.7", "composer", "install", "--no-dev", "--ignore-platform-reqs", "--no-interaction", "--no-progress", "--no-suggest", "--prefer-dist", "--optimize-autoloader"
                 },
                 function (code, success, error)
                     if code == 0 then
@@ -29,32 +29,32 @@ return vim.tbl_deep_extend(
         end,
         dependencies = {
             vim.tbl_deep_extend(
-            "force",
-            {
-                "gbprod/phpactor.nvim",
-                enabled = false,
-                tag = "v1.0.1",
-                config = function()
-                    require("phpactor").setup({
-                        install = {
-                            path = vim.fn.stdpath("data") .. "/lazy/phpactor",
-                            branch = "2023.09.24.0",
-                            bin = vim.fn.stdpath("data") .. "/lazy/phpactor/bin/phpactor",
-                            php_bin = "php",
-                            composer_bin = "composer2",
-                            git_bin = "git",
-                            check_on_startup = "none",
-                        },
-                        lspconfig = {
-                            enabled = false,
-                            options = {},
-                        },
-                    })
-                end
-            },
-            pelyib.config.phpactor
+                "force",
+                {
+                    "gbprod/phpactor.nvim",
+                    enabled = false,
+                    tag = "v1.0.1",
+                    config = function()
+                        require("phpactor").setup({
+                            install = {
+                                path = vim.fn.stdpath("data") .. "/lazy/phpactor",
+                                branch = "2023.09.24.0",
+                                bin = vim.fn.stdpath("data") .. "/lazy/phpactor/bin/phpactor",
+                                php_bin = "php",
+                                composer_bin = "composer2",
+                                git_bin = "git",
+                                check_on_startup = "none",
+                            },
+                            lspconfig = {
+                                enabled = false,
+                                options = {},
+                            },
+                        })
+                    end
+                },
+                pluginconf.phpactor or {}
             )
         },
     },
-    pelyib.config.phpactor
+    pluginconf.phpactor or {}
 )