+++ vim/dotfiles/.config/nvim/lua/plugins/toggleterm.lua
@@ -1,4 +1,4 @@
-local pelyib = vim.g.pelyib.pluginconfig
+local pluginconf = require('pelyib.pluginconf').config.patched
 
 return vim.tbl_deep_extend(
     "force",
@@ -15,5 +15,5 @@ return vim.tbl_deep_extend(
             })
         end
     },
-    pelyib.config.toggleterm
+    pluginconf.toggleterm or {}
 )