+++ vim/dotfiles/.config/nvim/lua/plugins/ufo.lua
@@ -1,4 +1,4 @@
-local pelyib = vim.g.pelyib.pluginconfig
+local pluginconf = require('pelyib.pluginconf').config.patched
 
 return vim.tbl_deep_extend(
     "force",
@@ -31,5 +31,5 @@ return vim.tbl_deep_extend(
             require('ufo').setup()
         end
     },
-    pelyib.config.ufo
+    pluginconf.ufo or {}
 )