+++ vim/test-files/js/src/index.js
@@ -1,0 +1,5 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import Title from "./title.js";
+
+ReactDOM.render(<Title />, document.getElementById("app"));