+++ vim/test-files/php-pest/src/App.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Neovim\Pest;
+
+class App
+{
+    public function run(): string
+    {
+        return 'Hello, World!';
+    }
+}